+++ Examples/Messenger-Shared/MessageViewController.m
@@ -522,9 +522,9 @@
 #endif
 }
 
-- (BOOL)shouldProcessTextForAutoCompletion:(NSString *)text
+- (BOOL)shouldProcessTextForAutoCompletion
 {
-    return [super shouldProcessTextForAutoCompletion:text];
+    return [super shouldProcessTextForAutoCompletion];
 }
 
 - (BOOL)shouldDisableTypingSuggestionForAutoCompletion