+++ Examples/Messenger-Swift/MessageViewController.swift
@@ -474,7 +474,7 @@ extension MessageViewController {
         }
     }
     
-    override func shouldProcessText(forAutoCompletion text: String) -> Bool {
+    override func shouldProcessTextForAutoCompletion() -> Bool {
         return true
     }