+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Change Log
 
+## Version 1.9.6
+
+##### Hot Fixes & Enhancements:
+- Fixed the text input not being interactive on iOS. Making `SLKTextInputbar` a UIView subclass instead of `UIToolbar`.
+
+##### Deprecation:
+- Deprecated `-shouldProcessTextForAutoCompletion:` in favor of `-shouldProcessTextForAutoCompletion`
+
+
 ## [Version 1.9.5](https://github.com/slackhq/SlackTextViewController/releases/tag/v1.9.5)
 
 ##### Features: